# fine-grain-logger: use memory order acquire and release