--- a/source/common/common/fine_grain_logger.h
+++ b/source/common/common/fine_grain_logger.h
@@ -197,10 +197,10 @@ FineGrainLogContext& getFineGrainLogContext();
 #define FINE_GRAIN_LOGGER()                                                                        \
   ([]() -> spdlog::logger* {                                                                       \
     static std::atomic<spdlog::logger*> flogger{nullptr};                                          \
-    spdlog::logger* local_flogger = flogger.load(std::memory_order_relaxed);                       \
+    spdlog::logger* local_flogger = flogger.load(std::memory_order_acquire);                       \
     if (!local_flogger) {                                                                          \
       ::Envoy::getFineGrainLogContext().initFineGrainLogger(__FILE__, flogger);                    \
-      local_flogger = flogger.load(std::memory_order_relaxed);                                     \
+      local_flogger = flogger.load(std::memory_order_acquire);                                     \
     }                                                                                              \
     return local_flogger;                                                                          \
   }())