@abeyad 2025-04-30
> cc @abeyad, this seems a consistent failure but I am not sure why for Mobile. Do you have any insights, thanks!
> 
> ```
> ERROR: /root/.cache/bazel/_bazel_root/bb3efe6b13248f746491cbfc24ae2efe/external/envoy/source/common/http/BUILD:138:17: Compiling source/common/http/codes.cc failed: (Exit 34): FAILED_PRECONDITION: Missing 1 digests
> java.io.IOException: io.grpc.StatusRuntimeException: FAILED_PRECONDITION: Missing 1 digests
> 	at com.google.devtools.build.lib.remote.GrpcRemoteExecutor.executeRemotely(GrpcRemoteExecutor.java:241)
> 	at com.google.devtools.build.lib.remote.RemoteExecutionService.executeRemotely(RemoteExecutionService.java:1906)
> ```

@botengyao I just reached out to the EngFlow team again, I think we'll have to open a PagerDuty issue.  I'll sync with you offline about this.